@@ -22,9 +22,8 @@ import './getNodeRenderers.css';
 
 export default function getNodeRenderers(findMatches: Set<string>, viewModifiers: Map<string, number>) {
   function vectorBorder(lv: TLayoutVertex<TDdgVertex>) {
-    // eslint-disable-next-line no-bitwise
     const isHovered = (viewModifiers.get(lv.vertex.key) || 0) & EViewModifier.Hovered;
-    // eslint-disable-next-line no-bitwise
+
     const isPathHovered = (viewModifiers.get(lv.vertex.key) || 0) & EViewModifier.PathHovered;
     const className = cx('DdgNode--VectorBorder', {
       'is-findMatch': findMatches.has(lv.vertex.key),
