@@ -133,9 +133,8 @@ describe('DeepDependencyGraph/url', () => {
 
     it('handles absent values', () => {
       ['end', 'hash', 'operation', 'service', 'start', 'visEncoding'].forEach(param => {
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
         const { [param]: unused, ...rest } = expectedParams;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
         const { [param]: alsoUnused, ...rv } = acceptableParams;
         parseSpy.mockReturnValue(rv);
         expect(getUrlState(getSearch())).toEqual(rest);
@@ -143,9 +142,8 @@ describe('DeepDependencyGraph/url', () => {
     });
 
     it("defaults `density` to 'ppe'", () => {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
       const { density: unused, ...rest } = expectedParams;
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+
       const { density: alsoUnused, ...rv } = acceptableParams;
       parseSpy.mockReturnValue(rv);
       expect(getUrlState(getSearch())).toEqual({ ...rest, density: 'ppe' });
