@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/* eslint-disable react/require-default-props */
-
 import React, { Component } from 'react';
 import { Col, Row, Tabs } from 'antd';
 import PropTypes from 'prop-types';
@@ -178,9 +176,9 @@ export class SearchTracePageImpl extends Component {
 }
 SearchTracePageImpl.propTypes = {
   isHomepage: PropTypes.bool,
-  // eslint-disable-next-line react/forbid-prop-types
+
   traceResultsToDownload: PropTypes.array,
-  // eslint-disable-next-line react/forbid-prop-types
+
   diffCohort: PropTypes.array,
   cohortAddTrace: PropTypes.func,
   cohortRemoveTrace: PropTypes.func,
