@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/* eslint-disable import/first */
 jest.mock('./scroll-page');
 
 import { scrollBy, scrollTo } from './scroll-page';
@@ -74,15 +73,14 @@ describe('ScrollManager', () => {
     });
 
     it('is a noop if an invalid rowPosition is returned by the accessors', () => {
-      // eslint-disable-next-line no-console
       const oldWarn = console.warn;
-      // eslint-disable-next-line no-console
+
       console.warn = () => {};
       manager._scrollPast(null, null);
       expect(accessors.getRowPosition.mock.calls.length).toBe(1);
       expect(accessors.getViewHeight.mock.calls.length).toBe(0);
       expect(scrollTo.mock.calls.length).toBe(0);
-      // eslint-disable-next-line no-console
+
       console.warn = oldWarn;
     });
 
