@@ -28,7 +28,7 @@ export default function TextList(props: TextListProps) {
         {data.map((row, i) => {
           return (
             // `i` is necessary in the key because row.key can repeat
-            // eslint-disable-next-line react/no-array-index-key
+
             <li key={`${i}`}>{row}</li>
           );
         })}
