@@ -171,7 +171,7 @@ describe('deepDependencyGraph reducers', () => {
     const multiPayload = {
       ...meta.query,
       visibilityIndices,
-      viewModifier: EViewModifier.Emphasized | EViewModifier.Selected, // eslint-disable-line no-bitwise
+      viewModifier: EViewModifier.Emphasized | EViewModifier.Selected,
     };
     const multiViewModifierMap = new Map();
     visibilityIndices.forEach(idx => multiViewModifierMap.set(idx, multiPayload.viewModifier));
@@ -242,7 +242,6 @@ describe('deepDependencyGraph reducers', () => {
           viewModifiers: new Map(),
         };
 
-        /* eslint-disable @typescript-eslint/no-unused-vars */
         const { operation: _op, ...emphasizedPayloadWithoutOp } = emphasizedPayload;
         const newState = addViewModifier(operationlessDoneState, emphasizedPayloadWithoutOp);
         const expected = _cloneDeep(operationlessDoneState);
