@@ -45,13 +45,13 @@ export function addViewModifier(state: TDdgState, payload: TDdgAddViewModifierPa
   const key = getStateEntryKey(payload);
   const stateEntry: TDdgStateEntry | void = state[key];
   if (!stateEntry || stateEntry.state !== fetchedState.DONE) {
-    console.warn('Cannot set view modifiers for unloaded Deep Dependency Graph'); // eslint-disable-line no-console
+    console.warn('Cannot set view modifiers for unloaded Deep Dependency Graph');
     return state;
   }
 
   const viewModifiers = new Map((stateEntry as IDoneState).viewModifiers);
   visibilityIndices.forEach(idx => {
-    viewModifiers.set(idx, (viewModifiers.get(idx) || 0) | viewModifier); // eslint-disable-line no-bitwise
+    viewModifiers.set(idx, (viewModifiers.get(idx) || 0) | viewModifier);
   });
 
   return {
@@ -68,7 +68,7 @@ export function viewModifierRemoval(state: TDdgState, payload: TDdgViewModifierR
   const key = getStateEntryKey(payload);
   const stateEntry: TDdgStateEntry | void = state[key];
   if (!stateEntry || stateEntry.state !== fetchedState.DONE) {
-    console.warn('Cannot change view modifiers for unloaded Deep Dependency Graph'); // eslint-disable-line no-console
+    console.warn('Cannot change view modifiers for unloaded Deep Dependency Graph');
     return state;
   }
 
@@ -76,9 +76,7 @@ export function viewModifierRemoval(state: TDdgState, payload: TDdgViewModifierR
   const indicesToUpdate = visibilityIndices || Array.from(viewModifiers.keys());
 
   indicesToUpdate.forEach(idx => {
-    const newValue = viewModifier
-      ? (viewModifiers.get(idx) || 0) & ~viewModifier // eslint-disable-line no-bitwise
-      : EViewModifier.None;
+    const newValue = viewModifier ? (viewModifiers.get(idx) || 0) & ~viewModifier : EViewModifier.None;
 
     if (newValue === EViewModifier.None) {
       viewModifiers.delete(idx);
