@@ -74,7 +74,7 @@ describe('google analytics tracking', () => {
       ]);
     });
 
-    it('check init function (no cookies)', () => {
+    it('check init function (with cookies)', () => {
       document.cookie = 'page=1;';
       tracking.init();
       expect(window.dataLayer).toEqual([
@@ -204,9 +204,7 @@ describe('google analytics tracking', () => {
       );
     });
 
-    /* eslint-disable no-console */
     it('isDebugMode = true', () => {
-      // eslint-disable-next-line no-import-assign
       console.log = jest.fn();
 
       trackingDebug.init();
