@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import * as React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies
+
 import { render } from 'react-dom';
 
 import { LayoutManager } from '../../src';
